nRF24_Wireless_Node description
This Arduino sketch is for the Wireless Flex Node board by ForceTronics.com. It creates a router or end device for a network of nRF24L01 transceivers using the RF24Network library TMRh20 version which is also available on GitHub. Note that this sketch uses the WSNode library that is also found here.